requirements.txt
requests==2.31.0
opencv-python==4.8.1.78
moviepy==1.0.3
Pillow==10.0.1
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2 